Final commit

+++ mvc/Controller/commands.php
@@ -9,20 +9,16 @@
 
 include 'mvc/Controller/orders.php';
 
-function XtoHome() {
-    
-    header('Location: ./?view=home');
-    exit;
-}
 
 function Xlogin() {
+    
+    redirectUser();
     
     if (isset($_POST['username'], $_POST['password']) && !empty($_POST['username']) && !empty($_POST['password'])) {
         if (User::verifyUser($_POST['username'], $_POST['password'])) {
             $_SESSION[USER] = $_POST['username'];
-            $_SESSION[TOKEN] = bin2hex(random_bytes(32));
-            $_SESSION['t2'] = bin2hex(openssl_random_pseudo_bytes(32));
-            header("Location: ./?view=hidden");
+            $_SESSION[TOKEN] = bin2hex(random_bytes(TOKEN_LEN));
+            headerLocationView('home');
             exit;
         }
         return 'Falsche Anmeldedaten';
@@ -32,57 +28,14 @@ function Xlogin() {
     }
 }
 
-function Xregister() {
-    if (isset($_POST['user'], $_POST['password']) && !empty($_POST['user']) && !empty($_POST['password']))
-    {
-        if(filter_var($_POST['user'], FILTER_VALIDATE_EMAIL)) {
-            if(preg_match('/[a-zA-Z0-9]{8,}/', $_POST['password'])) {
-                if(preg_match('/[0-9]/', $_POST['password']) == 1) {
-                    if(preg_match('/[$\_\.\-\+]/', $_POST['password'])) {
-                        if(preg_match('/[a-zA-Z]/', $_POST['password'])) {
-                            if (User::addUser($_POST['user'], $_POST['password'])) {
-                                if (User::verifyUser($_POST['user'], $_POST['password'])) {
-                                    $_SESSION[USER] = $_POST['user'];
-                                    header("Location: ./?view=hidden");
-                                    exit;
-                                }
-                                return 'Es ist ein Fehler aufgetreten';
-                            }
-                            return 'Benutzer existierts bereits';
-                        }
-                        return 'Mindestens ein Gross-/Kleinbuchstabe';
-                    }
-                    return 'Mindestens 1 Sonderzeichen';
-                }
-                return 'Mindestens 1 Zahl';
-            }
-            return 'Mindestens 8 Zeichen';
-        }
-        return 'Die E-Mail ist ungültig';
-    }
-    return 'Fülle alle Felder aus';
-}
-
-function XcreateGallery()
-{
-    if (isset($_POST['name'], $_POST['description']))
-    {
-        if (empty($_POST['name'])) {
-            return 'Gib einen Namen ein';
-        }
-        
-        $u = User::getUserByEmail($_SESSION[USER]);
-          
-        $g = new Gallery();
-        $g->addGallery(intval($u->getIdUser()), $_POST['name'], $_POST['description']);
-        header("Location: ");
-      }
-  }
-
 function XuploadImage(){
     
+    redirectGuest();
     
     if (isset($_POST['title'], $_POST['tags'], $_POST['galleryid']) && !empty($_POST['title']) && $_FILES['picture']['name'] != "") {
+        
+        if (!in_array(Gallery::getGalleryById($_POST['galleryid']), Gallery::getGalleriesByUserEmail($_SESSION[USER])))
+            return 'Diese Galerie gehört nicht dir';
         
         $file_size = $_FILES['picture']['size'];
         
@@ -92,6 +45,7 @@ function XuploadImage(){
             if ($file_size < 4000000) {
                 Picture::addPicture($_POST['galleryid'], $_POST['tags'], $_POST['title']);
                 //header("Location: ./?view=overview");
+                headerLocationView('gallery&id='. htmlentities($_POST['galleryid']));
             }
             return 'Das Bild ist zu gross';
         
@@ -105,11 +59,16 @@ function XuploadImage(){
 
 function XdeleteGallery()
 {
+    redirectGuest();
+    
     if (!isset($_POST['pw'], $_POST['gid']))
         return "Bitte gib dein Passwort ein";
     
     if (!User::verifyUser($_SESSION[USER], $_POST['pw']))
         return "Passwort ungültig";
+    
+    if (!in_array(Gallery::getGalleryById($_POST['gid']), Gallery::getGalleriesByUserEmail($_SESSION[USER])))
+        return 'Diese Galerie gehört nicht dir';
     
     $gallery = Gallery::getGalleryById($_POST['gid']);
     
@@ -123,8 +82,21 @@ function XdeleteGallery()
 
 function XremoveTagFromPic() {
     
+    redirectGuest();
+    
     if (!isset($_POST['pid'], $_POST['tagName']))
         return "Bitte wähle einen Tag zum entfernen aus";
+    
+    $userIsOwner = FALSE;
+    
+    foreach (Gallery::getGalleriesByUserEmail($_SESSION[USER]) as $gallery)
+    {
+        if (in_array(Picture::getPictureById($_POST['pid']), Picture::getPicturesFromGallery($gallery->getId())))
+            $userIsOwner = TRUE;
+    }
+    
+    if (!$userIsOwner)
+        return 'Das Bild gehört nicht dir';
     
     $picture = Picture::getPictureById($_POST['pid']);
     $tag = Tag::getTagByName($_POST['tagName']);
@@ -134,8 +106,22 @@ function XremoveTagFromPic() {
 }
 
 function XaddTagToPic() {
+    redirectGuest();
+    
     if (!isset($_POST['pid'], $_POST['tagName']))
         return "Bitte wähle einen Tag zum entfernen aus";
+    
+    $userIsOwner = FALSE;
+    
+    foreach (Gallery::getGalleriesByUserEmail($_SESSION[USER]) as $gallery)
+    {
+        if (in_array(Picture::getPictureById($_POST['pid']), Picture::getPicturesFromGallery($gallery->getId())))
+            $userIsOwner = TRUE;
+    }
+    
+    if (!$userIsOwner)
+        return 'Das Bild gehört nicht dir';
+    
     
     $picture = Picture::getPictureById($_POST['pid']);
     $picture->addTag($_POST['tagName']);
@@ -143,19 +129,36 @@ function XaddTagToPic() {
 }
 
 function Xsearch() {
-    if (!isset($_POST['search']))
+    
+    redirectGuest();
+    if (!isset($_REQUEST['search']))
         return "Bitte gebe einen Suchbegriff ein";
     
-    return $tags = Tag::searchPictures($_POST['search']);
+    return Tag::searchPicturesByUser($_REQUEST['search'], $_SESSION[USER]) ?: 'Es wurden keine Bilder mit diesem Tag gefunden';
 }
 
 function XdeletePicture()
 {
+    
+    redirectGuest();
+    
     if (!isset($_POST['pictureId']))
         return 'Ungültige ID';
     
     if (is_nan($_POST['pictureId']))
         return 'Ungültige ID';
+    
+    $userIsOwner = FALSE;
+    
+    foreach (Gallery::getGalleriesByUserEmail($_SESSION[USER]) as $gallery)
+    {
+        if (in_array(Picture::getPictureById($_POST['pictureIdw']), Picture::getPicturesFromGallery($gallery->getId())))
+            $userIsOwner = TRUE;
+    }
+    
+    if (!$userIsOwner)
+        return 'Das Bild gehört nicht dir';
+    
     
     Picture::deletePictureById($_POST['pictureId']);
